Remove obsolete Hashid class
The POJO has been implemented to simplify the handling of the binding
between the numbers and their hash value, but users should implement the
simple POJO on their own instead of forcing them to use this approach.

The public API methods now returning the primitive values which have
been hold by the Hashid class.

GH-9