Remove overloaded constructors
The overloaded constructors are not necessary since the Hashids.Builder
class provides methods for a accurate instance configuration. It allows
a finer grained building since it doesn't depends on parameter ordering.
The parameterless constructor is still available for brevity and the
functionality of all other removed constructors can be achieved with
the builder.

GH-9