Add initial gitbook documentation
This adds the initial documentation in form of a gitbook.
The book can be build by running the NPM script "docs:build",
for simplified development with live reload the NPM script "docs:serve"
can be used.

The book will be deployed as GitHub Page ("gh-pages" branch) including
the JavaDoc.

GH-11