Redesign project documentations
This commit contains a redeign of the project documentations to adapt to
the new gitbook docs. The changelog has been updated to match the new
layout and style and the README now matches the gitbook introduction and
also contains information on how to use the docs to get started and how
to use the contributing guide.

GH-10