Migrate to MIT license
Currently the project code is licensed under the Apache 2.0 (1) license.
This often causes problems when developers want to use the project or
code parts of it in another project (mainly code editors and UI
frameworks) licensed )nder a less restricted license.

The project and all of its port projects should migrate to the MIT license
(2) which is more open, unrestricted and the most used license for open
source projects (Facebook recently also re-licensed React (3))

References:

  (1) http://www.apache.org/licenses/LICENSE-2.0
  (2) https://choosealicense.com/licenses/mit/
  (3) https://code.facebook.com/posts/300798627056246/relicensing-react-jest-flow-and-immutable-js/

GH-31